US 2.7 #755: added tests checking member credit creation on subscription end date change

+++ tapir/payments/services/member_credit_creator.py
@@ -48,7 +48,7 @@ class MemberCreditCreator:
         cls,
         member: Member,
         actor: TapirUser,
-        amount_to_credit: float,
+        amount_to_credit: Decimal,
         reference_date: datetime.date,
         comment: str,
     ):
